started getting ready for ansible playbook

+++ base-install.sh
@@ -20,47 +20,50 @@ cfdisk "/dev/$DRIVE"
 # Formatting the drives
 echo "Formatting the drives:"
 DISKS_LIST=(`eval $DISKS`)
-echo "Did you make a swap partition? (Y/n)"
-read IS_SWAP
-if [[ $IS_SWAP =~ (yes)|(y)|(Y) ]]; then
-  echo "Which partition is the swap partition?"
-  echo "--------------------------------------"
-  echo "${DISKS_LIST[*]}"
-  read SWAP
-
-  mkswap "/dev/$SWAP"
-  if [[ $? != 0 ]]; then
-    echo "mkswap failed"
-    exit 1
-  fi
-fi
-
-if [[ $BIOS_TYPE == "uefi" ]]; then
-  echo "Did you make an efi partition? (Y/n)"
-  read IS_EFI
-  if [[ $IS_EFI =~ (yes)|(y)|(Y) ]]; then
-    echo "Which partition is the efi partition?"
+IS_FORMATTED=(`ls /mnt/`)
+if [[ ${IS_FORMATTED[@]} == 0 ]]; then
+  echo "Did you make a swap partition? (Y/n)"
+  read IS_SWAP
+  if [[ $IS_SWAP =~ (yes)|(y)|(Y) ]]; then
+    echo "Which partition is the swap partition?"
     echo "--------------------------------------"
     echo "${DISKS_LIST[*]}"
-    read EFI
+    read SWAP
 
-    mkfs.ext2 "/dev/$EFI"
+    mkswap "/dev/$SWAP"
     if [[ $? != 0 ]]; then
-      echo "mkfs failed"
+      echo "mkswap failed"
       exit 1
     fi
   fi
-fi
 
-echo "Which partition is your root partition?"
-echo "--------------------------------------"
-echo "${DISKS_LIST[*]}"
-read ROOT
+  if [[ $BIOS_TYPE == "uefi" ]]; then
+    echo "Did you make an efi partition? (Y/n)"
+    read IS_EFI
+    if [[ $IS_EFI =~ (yes)|(y)|(Y) ]]; then
+      echo "Which partition is the efi partition?"
+      echo "--------------------------------------"
+      echo "${DISKS_LIST[*]}"
+      read EFI
 
-mkfs.ext4 "/dev/$ROOT"
-if [[ $? != 0 ]]; then
-  echo "mkfs failed"
-  exit 1
+      mkfs.ext2 "/dev/$EFI"
+      if [[ $? != 0 ]]; then
+        echo "mkfs failed"
+        exit 1
+      fi
+    fi
+  fi
+
+  echo "Which partition is your root partition?"
+  echo "--------------------------------------"
+  echo "${DISKS_LIST[*]}"
+  read ROOT
+
+  mkfs.ext4 "/dev/$ROOT"
+  if [[ $? != 0 ]]; then
+    echo "mkfs failed"
+    exit 1
+  fi
 fi
 
 # Mount file systems
@@ -72,7 +75,7 @@ fi
 
 # Install Base System
 echo "Installing the base linux system:"
-pacstrap /mnt base linux linux-firmware
+pacstrap /mnt base linux linux-firmware ansible
 
 # Generate fstab
 echo "Generating fstab file:"
